+++ CHANGELOG.md
@@ -1,8 +1,16 @@
+v3.0.4
+------
+- Updated Slang to 0.10.24
+- Added an option to create a `Program` from a string
+- Added `CopyContext::updateSubresourceData()` which allows updating a region of a subresource
+- Added `Program::Desc` has a new function - `setShaderModel()`. It allows the user to request shader-model 6.x, which will use dxcompiler instead of FXC
+
 v3.0.3
 ------
 - Added FXAA as an effect
 - Support programmable sample position - `Fbo::setSamplePositions()` (DX only)
 - Added RenderContext::resolveResource() and RenderContext::resolveSubresource() MSAA resolve functions
+- Added support for setting shading model through fscene files and load flags. Also editable in Scene Editor
 
 v3.0.2
 ------